Split eventEntries class to its own header, made tracelib header compilable with C
